🎨 Header를 컴포넌트로 만들었습니다.
